Prevent overflow in graph node page HasMore (#3875)
* fix: prevent graph paging HasMore overflow

* test: graph paging extreme page boundary

+++ ArchLucid.Core/Persistence/Graph/GraphSnapshotPagination.cs
@@ -27,7 +27,7 @@ public static class GraphSnapshotPagination
             Page = safePage,
             PageSize = safeSize,
             TotalNodes = total,
-            HasMore = safePage * safeSize < total,
+            HasMore = (long)safePage * safeSize < total,
             Nodes = slice,
             Edges = edges
         };
